Fix malloc and memcpy. Fix CLZ for no-xlp. Use C version of memcpy.

+++ stdlib/malloc.c
@@ -85,7 +85,7 @@ void *malloc (size_t size)
       if (IS_FREE(block))
         {
           rem_size = block->size - alloc_size;
-          if (rem_size >= 0)
+          if ((long)rem_size >= 0)
             {
               /* Split this chunk */
               if (rem_size >= MALLOC_MINCHUNK)
